添加默认时间

+++ roco-oa-web-e/src/composables/demand/demandPeopleComposable.ts
@@ -39,6 +39,9 @@ export default function demandPeopleComposable() {
       return demandPerson.userId;
     });
     await saveDemandPeopleAPI(demandPeople.value);
+    if (!deleteDemandPeopleIds.length) {
+      return;
+    }
     await deleteDemandPeopleAPI(deleteDemandPeopleIds);
   };
